MySQL driver: full nothrow-first refactor with static factory pattern, result set registry, and dead try/catch elimination
- MySQLDBDriver: double-checked locking (atomic<bool> + mutex) for library init; new cleanup() static method
- MySQLDBConnection: PrivateCtorTag pattern, nothrow constructor with m_initFailed/m_initError, result set registry (m_activeResultSets) with two-phase close, getMySQLNativeHandle()/getConnectionMutex() for child access, m_resetting anti-deadlock flag
- MySQLDBPreparedStatement: weak_ptr<MySQLDBConnection> replaces weak_ptr<MYSQL>, private nothrow constructor, create(std::nothrow_t) via new, m_closed atomic flag
- MySQLDBResultSet: private nothrow constructor, weak_ptr<MySQLDBConnection> for lifecycle, notifyConnClosing(std::nothrow_t), [[nodiscard]] on all nothrow methods
- mysql_internal.hpp: MySQLConnectionLock RAII helper, MYSQL_CONNECTION_LOCK_OR_RETURN macro, MYSQL_DEBUG macro
- Dead try/catch eliminated across all MySQL .cpp files
- Updated all project documentation (CHANGELOG, TODO, README, memory-bank, docs)
- 26 files changed, +1683/-2231 lines (net reduction of ~548 lines)

+++ libs/cpp_dbc/src/pool/connection_pool.cpp
@@ -318,13 +318,15 @@ namespace cpp_dbc
                         continue;
                     }
 
-                    auto idleTime = std::chrono::duration_cast<std::chrono::milliseconds>(
-                                        now - pooledConn->getLastUsedTime(std::nothrow))
-                                        .count();
+                    auto idleTime = static_cast<size_t>(
+                        std::chrono::duration_cast<std::chrono::milliseconds>(
+                            now - pooledConn->getLastUsedTime(std::nothrow))
+                            .count());
 
-                    auto lifeTime = std::chrono::duration_cast<std::chrono::milliseconds>(
-                                        now - pooledConn->getCreationTime(std::nothrow))
-                                        .count();
+                    auto lifeTime = static_cast<size_t>(
+                        std::chrono::duration_cast<std::chrono::milliseconds>(
+                            now - pooledConn->getCreationTime(std::nothrow))
+                            .count());
 
                     bool expired = (idleTime > m_idleTimeoutMillis) ||
                                    (lifeTime > m_maxLifetimeMillis);
@@ -795,11 +797,11 @@ namespace cpp_dbc
             {
                 auto lastUsed = result->getLastUsedTime(std::nothrow);
                 auto now = steady_clock::now();
-                auto timeSinceLastUse = duration_cast<milliseconds>(now - lastUsed).count();
+                auto timeSinceLastUse = static_cast<size_t>(duration_cast<milliseconds>(now - lastUsed).count());
 
                 if (timeSinceLastUse > m_validationTimeoutMillis)
                 {
-                    CP_DEBUG("DBConnectionPoolBase::acquireConnection - VALIDATING (timeSinceLastUse=%lld ms > %zu ms)", (long long)timeSinceLastUse, m_validationTimeoutMillis);
+                    CP_DEBUG("DBConnectionPoolBase::acquireConnection - VALIDATING (timeSinceLastUse=%zu ms > %zu ms)", timeSinceLastUse, m_validationTimeoutMillis);
                     auto valResult = validateConnection(std::nothrow, result->getUnderlyingConnection(std::nothrow));
                     bool validationPassed = valResult.has_value() && valResult.value();
